refactor(LottoTest): 로또 객체 테스트 코드 추가 및 리팩토링
